+++ GenericFanSite/Controllers/ForumController.cs
@@ -3,7 +3,6 @@ using Microsoft.AspNetCore.Mvc;
 using GenericFanSite.Data;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
-using static System.Net.Mime.MediaTypeNames;
 
 namespace GenericFanSite.Controllers
 {
@@ -17,7 +16,7 @@ namespace GenericFanSite.Controllers
             repo = r;
         }
         [HttpGet]
-        public async Task<IActionResult> Index(ForumSearchVM data)
+        public async Task<IActionResult> IndexAsync(ForumSearchVM data)
         {
             int countFromResults = data.Results;
             if (data.Results == 0)  //Default for number of forum posts displayed is 5.
@@ -100,69 +99,6 @@ namespace GenericFanSite.Controllers
                 }
             }
             return View(data);
-        }
-        [Authorize]
-        [HttpGet]
-        public async Task<IActionResult> ForumPostSingle(int forumPostId)
-        {
-            ForumPost data = repo.GetForumPostByIdAsync(forumPostId).Result;
-            return View(data);
-        }
-        [Authorize]
-        [HttpPost]
-        public async Task<IActionResult> ForumPostSingle(int forumPostId, Comment comment)
-        {
-            ForumPost data = repo.GetForumPostByIdAsync(forumPostId).Result;
-            if (!(comment.CommentText == null))
-            {
-                comment.CommentText = comment.CommentText.Trim();
-                comment.Date = DateTime.Now;
-                comment.User = await userManager.GetUserAsync(User);
-            }
-            else
-            {
-                return View(data);
-            }
-            ModelState.Remove(nameof(comment.User));
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    data.Comments.Add(comment);
-                    if (data != null && await repo.UpdateForumPostAsync(data) > 0)
-                    {
-                        return View(data);
-                    }
-                    else
-                    {
-                        ViewBag.RedText = "There was a really bad error saving the forum post.";
-                        return View();
-                    }
-                }
-                catch
-                {
-                    ViewBag.RedText = "An unknown error has occured.";
-                    return View(data);
-                }
-            }
-            return View(data);
-        }
-        [Authorize]
-        public IActionResult DeleteForumPost(int forumPostId)
-        {
-            if (repo.DeleteForumPost(forumPostId) > 0) {
-                return RedirectToAction("Index");
-            }
-            else
-            {
-                return RedirectToAction("ForumPostSingle", forumPostId);
-            }
-        }
-        [Authorize]
-        public IActionResult DeleteComment(int forumPostId, int commentId)
-        {
-            repo.DeleteComment(forumPostId, commentId);
-            return RedirectToAction("ForumPostSingle", "Forum", new { forumPostId = forumPostId });
         }
     }
 }
