+++ GenericFanSite/Controllers/HomeController.cs
@@ -7,32 +7,38 @@ namespace GenericFanSite.Controllers
     public class HomeController : Controller
     {
         private readonly ILogger<HomeController> _logger;
+
         public HomeController(ILogger<HomeController> logger)
         {
             _logger = logger;
         }
+
         public IActionResult Index()
         {
             return View();
         }
+
         public IActionResult History()
         {
             return View();
         }
         public IActionResult Stories()
         {
-            return RedirectToAction("Index", "Forum");
+            return View();
         }
+
         public IActionResult Privacy()
         {
             return View();
         }
+
         [HttpGet]
         public IActionResult Quiz()
         {
             Quiz model = new Quiz();
             return View(model);
         }
+
         [HttpPost]
         public IActionResult Quiz(string[] answers)
         {
@@ -46,6 +52,7 @@ namespace GenericFanSite.Controllers
             }
             return View(model);
         }
+
         [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
         public IActionResult Error()
         {
