+++ GenericFanSite/Data/AppDbContext.cs
@@ -4,7 +4,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace GenericFanSite.Data
 {
-    public class AppDbContext : IdentityDbContext<AppUser>
+    public class AppDbContext : IdentityDbContext
     {
         // constructor just calls the base class constructor
         public AppDbContext(
