+++ GenericFanSite/Data/FakeForumRepo.cs
@@ -26,19 +26,5 @@ namespace GenericFanSite.Data
             }
             return status;
         }
-        public async Task<int> UpdateForumPostAsync(ForumPost model)
-        {
-            return -1;
-        }
-        public int DeleteForumPost(int id)
-        {
-            /*context.Remove(id);
-            return context.SaveChanges();*/
-            throw new Exception();
-        }
-        public int DeleteComment(int forumPostId, int commentId)
-        {
-            throw new Exception();
-        }
     }
 }
