--- GenericFanSite/Views/Forum/ForumPostSingle.cshtml
@@ -1,58 +1,0 @@
-﻿@using Microsoft.AspNetCore.Identity
-@model ForumPost
-@inject SignInManager<AppUser> signInManager
-@{
-    ViewData["Title"] = Model.User.ToString() + "'s Story";
-}
-<a asp-controller="Home" asp-action="Index">Back to Home</a>
-<section class="text-center">
-    <h1>@ViewData["Title"]:</h1>
-</section>
-<hr />
-<article class="container-fluid mt-3">
-    <div class="mt-md-3 text-center row d-flex justify-content-center">
-        <div class="bg-dark-subtle col col-md-8 mb-3">
-            <!--Forum post should only stretch to full width on smaller screens.-->
-            <h3>@Model.Title</h3>
-            <p>@Model.Description</p>
-            <p>Year this story took place: @Model.Year</p>
-            <hr />
-            <p>@Model.Story</p>
-            <hr />
-            <p class="text-start float-start mb-0">Author: @Model.User.UserName &nbsp;</p>
-            <abbr class="text-sm-start text-start float-start float-sm-end mb-0" title="@Model.Date.ToString()">Posted on: @Model.Date.ToShortDateString()</abbr><!--Text should float to left and right sides except on small screens.-->
-            @if (User.Identity.Name == Model.User.UserName || User.IsInRole("Admin"))
-            {
-                <br />
-                <a class="text-danger text-sm-start text-start float-start float-sm-end" asp-controller="Forum" asp-action="DeleteForumPost" asp-route-forumPostId="@Model.ForumPostId">Delete Post</a>
-            }
-        </div>
-    </div>
-    <div class="mt-md-3 row d-flex justify-content-center">
-        <div class="bg-dark-subtle col col-md-8 mb-3">
-        <h3>Comments:</h3>
-        <form class="input-group mb-3" method="post">
-            <input type="text" class="form-control" placeholder="Add a comment" name="CommentText">
-            <div class="input-group-append">
-                <button class="btn btn-outline-secondary" type="submit">Submit</button>
-            </div>
-        </form>
-        @if (Model.Comments.Count > 0) {
-            foreach (Comment comment in Model.Comments) {
-                <hr class="my-1"/>
-                <p class="mb-1">@comment.User:</p>
-                <p class="mb-1">@comment.CommentText</p>
-                <abbr class="text-secondary" title="@comment.Date.ToString()">@comment.Date.ToShortDateString()</abbr>
-                @if (User.Identity.Name == comment.User.UserName || User.IsInRole("Admin"))
-                {
-                    <a class="text-danger text-sm-start text-start float-start float-sm-end" asp-controller="Forum" asp-action="DeleteComment" asp-route-forumPostId="@Model.ForumPostId" asp-route-commentId="@comment.CommentId">Delete Comment</a>
-                }
-            }
-        }
-        else {
-            <hr />
-            <p class="text-danger">There are no comments.</p>
-        }
-        </div>
-    </div>
-</article>
