+++ GenericFanSite/Views/Forum/Index.cshtml
@@ -5,7 +5,7 @@
     ViewData["Title"] = "All Stories";
 }
 @model ForumSearchVM
-<a asp-controller="Home" asp-action="Index">Back to Home</a>
+<br />
 <section class="text-center">
     <h1>@ViewData["Title"]:</h1>
 </section>
@@ -95,10 +95,8 @@
                     <hr />
                     <p>@forumPost.Story</p>
                     <hr />
-                    <p class="text-start float-start mb-0">Author: @forumPost.User.UserName &nbsp;</p>
-                    <p class="text-sm-start text-start float-start float-sm-end mb-0">Posted on: @forumPost.Date.ToShortDateString()</p><!--Text should float to left and right sides except on small screens.-->
-                    <br />
-                    <a class="" asp-controller="Forum" asp-action="ForumPostSingle" asp-route-forumPostId="@forumPost.ForumPostId">View Comments</a>
+                    <p class="text-start float-start">Author: @forumPost.User.UserName &nbsp;</p>
+                    <p class="text-sm-start text-start float-start float-sm-end">Posted on: @forumPost.Date.ToShortDateString()</p><!--Text should float to left and right sides except on small screens.-->
                 </div>
             </div>
         }
