+++ GenericFanSite/Views/Home/History.cshtml
@@ -2,22 +2,8 @@
     ViewData["Title"] = "History";
 }
 <!--TODO: Write text.-->
-<a asp-controller="Home" asp-action="Index">Back to Home</a>
-<section class="text-center">
+<a asp-area="" asp-controller="Home" asp-action="Index">Back to Home</a>
+<article class="text-center">
     <h1>@ViewData["Title"]</h1>
     <p>A history of Mark Foster and his band.</p>
-</section>
-<article class="container-fluid">
-    <p>
-        Lead singer and musician Mark Foster founded the band Foster the People in 2009.
-        The band has had a few different members come and go since its founding, with musician and pianist Isom Innis being the longest-standing, who still goes on tour to this day, since his introduction in 2010.
-    </p>
-    <h4>Albums released:</h4>
-    <ul>
-        <li><a href="https://open.spotify.com/album/2wu3EA8IvohnL7Ik88Usl9">Torches</a></li>
-        <li><a href="https://open.spotify.com/album/22cFcAQkydpTzeSKQZEKv0">Supermodel</a></li>
-        <li><a href="https://open.spotify.com/album/5lFvZh6pCTJzr9UStebyCF">Sacred Hearts Club</a></li>
-        <li><a href="https://open.spotify.com/album/4UaPlMSS15FOS4A858WULM">In the Darkest of Nights, Let the Birds Sing</a></li>
-        <li><a href="https://open.spotify.com/album/27ynHS80OjICdw3qLNMgQP">Paradise State of Mind</a></li>
-    </ul>
 </article>
