+++ GenericFanSite/Views/Home/Stories.cshtml
@@ -1,0 +1,18 @@
+﻿@{
+    ViewData["Title"] = "Stories";
+}
+<a asp-area="" asp-controller="Home" asp-action="Index">Back to Home</a>
+<article class="text-center">
+    <h1>@ViewData["Title"]</h1>
+    <p>Stories from fans about their favorite moments with the band.</p>
+</article>
+<article>
+    <ul>
+        <li>
+            <a asp-controller="Forum" asp-action="Index">View all stories</a>
+        </li>
+        <li>
+            <a asp-controller="Forum" asp-action="ForumPostForm">Create a forum post</a>
+        </li>
+    </ul>
+</article>
