+++ GenericFanSite/Views/Shared/_Layout.cshtml
@@ -37,14 +37,8 @@
                                 <li><a class="dropdown-item text-white" asp-area="" asp-controller="Home" asp-action="Stories">Stories</a></li>
                             </ul>
                         </li>
-                        <li class="nav-item dropdown">
-                            <a class="nav-link dropdown-toggle text-dark" data-bs-toggle="dropdown" role="button" aria-expanded="false">Sources</a>
-                            <ul class="dropdown-menu rounded-0 border-light">
-                                <li><a class="dropdown-item" asp-area="" asp-controller="Sources" asp-action="Index">Sources</a></li>
-                                <li><hr class="dropdown-divider"></li>
-                                <li><a class="dropdown-item text-white" asp-area="" asp-controller="Sources" asp-action="FanSites">Fan Sites</a></li>
-                                <li><a class="dropdown-item text-white" asp-area="" asp-controller="Sources" asp-action="News">News</a></li>
-                            </ul>
+                        <li class="nav-item">
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Sources" asp-action="Index">Sources</a>
                         </li>
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-controller="Archive" asp-action="Index">Archive</a>
